Use opaque content material instead of vibrant sidebar material for EQ panes
The previous fix for the canvas/sidebar background mismatch in dark
mode gave the EQ canvas/readout/footer the same vibrant .sidebar
NSVisualEffectView material as the actual sidebar. That made the
content pane translucent too, which isn't how native split-view apps
look — Finder/Mail/Xcode keep the sidebar vibrant but the content pane
solid/opaque.

Switch to .contentBackground material with .withinWindow blending
(ContentMaterialBackground, renamed from SidebarMaterialBackground).
Still a real native material rather than a flat Color(nsColor:) fill,
so it still resolves correctly in dark mode, but it stays opaque like
a real content pane should.
