+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to iQualize will be documented in this file.
 ### Added
 - A native collapsible sidebar in the EQ window, toggled from the toolbar's new leftmost button, listing presets in Favorites/Built-in/Custom sections with a star button to toggle favorites. Selecting a row loads it through the same `DreamViewModel.loadPreset(id:)` path as the existing toolbar preset picker, so the unsaved-changes confirmation still applies. This is additive — the toolbar's preset-picker popup menu is unchanged and both remain side by side. Implemented via a new `NSSplitViewController` (`DreamSplitViewController`), replacing the window's plain `NSView` content view; collapsed by default, state and width persist across launches.
 - The window now uses `.fullSizeContentView` with a transparent title bar, so the sidebar's background extends up behind the traffic lights and the toolbar floats over it, Safari-style, instead of stopping at an opaque toolbar strip. The window title text is hidden so the sidebar-toggle button sits flush against the traffic lights (Xcode/Notes/Mail convention) rather than being pushed right by reserved title space. An `NSTrackingSeparatorToolbarItem` ties the toggle button to the sidebar/content divider, so it hugs the sidebar's right edge and moves with it when resized (Safari's "button lives inside the sidebar" behavior) instead of sitting in the fixed leading toolbar group.
-- The EQ canvas/readout/footer backgrounds now use the same native `NSVisualEffectView` sidebar material as the preset sidebar (`SidebarMaterialBackground`), replacing a flat `windowBackgroundColor` fill that matched in light mode but visibly diverged from the sidebar's vibrant material in dark mode. It also correctly dims when the window loses focus, matching the native sidebar list instead of always rendering as if key. `DreamTheme.bgWindow`/`bgToolbar`/`bgCanvas`/`bgTitlebar` are removed as unused. The sidebar's favorite-star accent color changed from a hardcoded yellow to the system accent color (`NSColor.controlAccentColor`).
+- The EQ canvas/readout/footer backgrounds now use a native `NSVisualEffectView` (`ContentMaterialBackground`, `.contentBackground` material) instead of a flat `windowBackgroundColor` fill that matched the sidebar's tone in light mode but visibly diverged from it in dark mode — the same class of solid content-pane background Finder/Mail/Xcode use next to a vibrant sidebar (the sidebar itself keeps its native vibrant `.sidebar` material; content panes stay opaque, not translucent). It also correctly dims when the window loses focus, matching the native sidebar list instead of always rendering as if key. `DreamTheme.bgWindow`/`bgToolbar`/`bgCanvas`/`bgTitlebar` are removed as unused. The sidebar's favorite-star accent color changed from a hardcoded yellow to the system accent color (`NSColor.controlAccentColor`).
 
 ## [0.55.0] - 2026-07-31
