+++ Sources/iQualize/DreamUI/EQReadoutGrid.swift
@@ -17,7 +17,7 @@ struct EQReadoutGrid: View {
         }
         .padding(.horizontal, 12)
         .padding(.vertical, 6)
-        .background(SidebarMaterialBackground())
+        .background(ContentMaterialBackground())
     }
 
     @ViewBuilder
